refactor: change messageId type from long to int across codebase (#214)
Telegram Bot API uses 32-bit int for message IDs (unique per chat, not globally).
This aligns all layers with the SDK native type, eliminating ~16 explicit
(int) casts at the Telegram.Bot boundary.

WIP: Migration generated but not yet reviewed or applied.

Changes across 100 files:
- Core: JobPayloads (FileScanJobPayload, FetchUserPhotoPayload)
- Data: All DTOs, views, and model snapshot
- ContentDetection: Models, repository interfaces and implementations
- Telegram: Models, ModerationIntents, services, repos, handlers
- BackgroundJobs: Removed (int) casts in FileScanJob, DeleteUserMessagesJob
- Main app: Blazor components, analytics models
- Tests: Unit, component, integration, and E2E tests updated

+++ TelegramGroupsAdmin.BackgroundJobs/Jobs/FileScanJob.cs
@@ -248,7 +248,7 @@ public class FileScanJob(
         try
         {
             // Delete the message containing infected file
-            await _messageService.DeleteAndMarkMessageAsync(payload.Chat.Id, (int)payload.MessageId, "file_scan_infected", cancellationToken);
+            await _messageService.DeleteAndMarkMessageAsync(payload.Chat.Id, payload.MessageId, "file_scan_infected", cancellationToken);
 
             _logger.LogWarning(
                 "Deleted infected file message {MessageId} from {Chat}",
