+++ TelegramGroupsAdmin.Data/Models/EnrichedMessageView.cs
@@ -85,7 +85,7 @@ public class EnrichedMessageView
     #region Message Columns (from messages table)
 
     [Column("message_id")]
-    public long MessageId { get; set; }
+    public int MessageId { get; set; }
 
     [Column("user_id")]
     public long UserId { get; set; }
@@ -127,7 +127,7 @@ public class EnrichedMessageView
     public string? DeletionSource { get; set; }
 
     [Column("reply_to_message_id")]
-    public long? ReplyToMessageId { get; set; }
+    public int? ReplyToMessageId { get; set; }
 
     [Column("media_type")]
     public MediaType? MediaType { get; set; }
